fmri patch

+++ lib/neuroimaging/fmri/__init__.py
@@ -92,9 +92,9 @@ class fMRISamplingGrid(SamplingGrid):
 
             def _map(x, fn=self.mapping.map, **keywords):
                 if len(x.shape) > 1:
-                    _x = N.zeros((x.shape[0]+1,) + x.shape[1:], N.Float)
+                    _x = N.zeros((x.shape[0]+1,) + x.shape[1:], N.float64)
                 else:
-                    _x = N.zeros((x.shape[0]+1,), N.Float)
+                    _x = N.zeros((x.shape[0]+1,), N.float64)
                 _x[0] = i
                 return fn(_x)
             W = Mapping(incoords, outcoords, _map)
